Only remove attachment if physical name is present
git-svn-id: https://subversion.assembla.com/svn/phpBB_Tracker/trunk@59 e70539a1-e9aa-4707-9e48-a9b25a7622e0

+++ includes/tracker/tracker_class.php
@@ -350,9 +350,13 @@ class tracker
 			WHERE attach_id = ' . $filedata['attach_id'];
 		$db->sql_query($sql);
 
-		$filename = basename($filedata['physical_filename']);
-
-		return @unlink($phpbb_root_path . $this->config['attachment_path'] . '/' . $filename);
+		if (isset($filedata['physical_filename']))
+		{
+			$filename = basename($filedata['physical_filename']);
+			return @unlink($phpbb_root_path . $this->config['attachment_path'] . '/' . $filename);
+		}
+		
+		return true;		
 	}
